third_party/blink/web_tests/editing/inserting/insert-table-in-paragraph-crash-expected.txt

PASS $("editable").innerHTML is "<table><tbody><tr><td>stats</td></tr></tbody></table>"
PASS successfullyParsed is true

TEST COMPLETE
